# docs/api/openapi.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Beyond Style UAE — AI Jewellery Designer API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/products/catalogue": {
   "get": {
    "description": "Owner catalogue items with stated starting prices (AED).",
    "operationId": "product_catalogue_api_products_catalogue_get",
    "parameters": [
     {
      "in": "query",
      "name": "family",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Family"
      }
     },
     {
      "in": "query",
      "name": "text_process",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Text Process"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Product Catalogue",
    "tags": [
     "products"
    ]
   }
  },
  "/api/products/catalogue/{code}": {
   "get": {
    "operationId": "product_catalogue_item_api_products_catalogue__code__get",
    "parameters": [
     {
      "in": "path",
      "name": "code",
      "required": true,
      "schema": {
       "title": "Code",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Product Catalogue Item",
    "tags": [
     "products"
    ]
   }
  },
...
  "/api/products/size-guide/bracelet": {
   "get": {
    "operationId": "bracelet_size_api_products_size_guide_bracelet_get",
    "parameters": [
     {
      "in": "query",
      "name": "wrist_cm",
      "required": true,
      "schema": {
       "title": "Wrist Cm",
       "type": "number"
      }
     },
     {
      "in": "query",
      "name": "fit",
      "required": false,
      "schema": {
       "default": "standard",
       "title": "Fit",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Bracelet Size",
    "tags": [
     "products"
    ]
   }
  },
...
 }
}